Consolidate stale-related utils in `oak_scan`
