+++ crates/oak_db/src/tests/test_db.rs
@@ -108,10 +108,6 @@ impl Db for TestDb {
         crate::db::package_by_name_query(self, name)
     }
 
-    fn package_by_url(&self, url: &UrlId) -> Option<crate::Package> {
-        crate::db::package_by_url_query(self, url)
-    }
-
     fn root_by_package(&self, pkg: crate::Package) -> Option<crate::Root> {
         crate::db::root_by_package_query(self, pkg)
     }
