+++ crates/oak_scan/src/lib.rs
@@ -24,6 +24,7 @@
 
 mod inputs;
 mod library;
+mod lookup;
 mod packages;
 mod stale;
